+++ CHANGELOG.md
@@ -11,6 +11,10 @@ The format is based on [Keep a Changelog][changelog] and this project adheres to
 
 - The `cat` command to decrypt and print a particular set of content URIs
 
+### Changed
+
+- The `preview` command to show what a `create` would do
+
 ## 10.0.0 - 2021-05-20
 
 ### Changed